+++ lib/oxidized/source/jsonfile.rb
@@ -26,11 +26,7 @@ module Oxidized
       end
 
       def load(*)
-        file = open_file
-        # open_file returns a File on the plain path but a String on the gpg
-        # path (see Source#open_file), so only call #read when it is supported.
-        contents = file.respond_to?(:read) ? file.read : file
-        data = JSON.parse(contents)
+        data = JSON.parse(open_file.read)
         data = string_navigate_object(data, @cfg.hosts_location) if @cfg.hosts_location?
 
         transform_json(data)